move wallet trial to new trials dir
