--- packages/lavamoat-browserify-perf/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>LavaMoat Examples</title>
-</head>
-<body>
-    <script src="/bundle.js" type="text/javascript"></script>
-</body>
-</html>
