--- packages/lavamoat-browserify-perf/perf.csv
@@ -1,8 +1,0 @@
-n, without,with,harden
-1,0.167778872,0.485361434,0.395091113
-50,0.22407727,0.65264229,0.598295674
-100,0.281073721,0.86863608,0.764584582
-250,0.442282479,1.421490801,1.28092578
-500,0.623277429,2.268732108,2.0771874
-750,0.839173755,3.17664137,2.773034923
-1000,0.998225104,3.985438183,3.529306458
