+++ packages/lavamoat-browserify-perf/trials/wallet/chart.html
@@ -70,14 +70,14 @@ function run () {
 }
 
 var csv = `
-n, without,with,harden
-1,0.167778872,0.485361434,0.395091113
-50,0.22407727,0.65264229,0.598295674
-100,0.281073721,0.86863608,0.764584582
-250,0.442282479,1.421490801,1.28092578
-500,0.623277429,2.268732108,2.0771874
-750,0.839173755,3.17664137,2.773034923
-1000,0.998225104,3.985438183,3.529306458
+n, node,bify,bify+lavamoat,bify+lavamoat w/ harden
+1,0.086893314,0.146961837,0.403807913,0.354210107
+50,0.082383619,0.220393655,0.611215764,0.53845047
+100,0.088283082,0.265781174,0.82052651,0.73275201
+250,0.102358306,0.415043594,1.310680593,1.174689173
+500,0.112837484,0.58082724,2.207336968,1.931934112
+750,0.133747079,0.771013143,3.019740428,2.679880295
+1000,0.154356264,0.927030139,3.912655944,3.44262978
 `
 
 </script>
