+++ packages/lavamoat-browserify-perf/trials/wallet/entry.js
@@ -1,5 +1,6 @@
 const { fromEtherCamp: brainwallet } = require('ethereumjs-wallet/thirdparty')
 
+// use `global.process` so it works correctly under node
 const nTimes = Number.parseInt(global.process.env.PERF_N || 5, 10)
 
 Array(nTimes).fill().forEach((_, index) => {
