+++ packages/lavamoat-browserify-perf/trials/wallet/lavamoat-config-harden.json
