+++ packages/lavamoat-browserify-perf/trials/wallet/lavamoat-config.json
