+++ packages/lavamoat-browserify-perf/trials/wallet/perf.csv
@@ -1,0 +1,8 @@
+n, node,bify,bify+lavamoat,bify+lavamoat w/ harden
+1,0.086893314,0.146961837,0.403807913,0.354210107
+50,0.082383619,0.220393655,0.611215764,0.53845047
+100,0.088283082,0.265781174,0.82052651,0.73275201
+250,0.102358306,0.415043594,1.310680593,1.174689173
+500,0.112837484,0.58082724,2.207336968,1.931934112
+750,0.133747079,0.771013143,3.019740428,2.679880295
+1000,0.154356264,0.927030139,3.912655944,3.44262978
